+++ src/FiapCloundGames.API/Application/Dtos/Identity/LoginResponse.cs
@@ -1,4 +1,7 @@
-﻿namespace FiapCloundGames.API.Application.Dtos.Usuario
+﻿using FiapCloundGames.API.Application.Dtos.Usuario;
+using FiapCloundGames.API.Domain.Enum;
+
+namespace FiapCloundGames.API.Application.Dtos.Identity
 {
     public class LoginResponse
     {
@@ -7,5 +10,6 @@
         public string Id { get; set; }
         public string Email { get; set; }
         public IEnumerable<ClaimResponse> Claims { get; set; }
+        public TipoUsuario PerfilUsuario { get; set; }
     }
 }
